refactor(algo) : policy gradient
