+++ omnisafe/configs/on-policy/PolicyGradient.yaml
@@ -57,6 +57,14 @@ defaults:
   target_kl: 0.02
   # The size of batch for policy update
   batch_size: 10000
+  # The number of parallel environments
+  num_envs: 8
+  # Whether to use standardized reward
+  reward_normalize: True
+  # Whether to use standardized cost
+  cost_normalize: True
+  # Whether to use standardized obs
+  obs_normalize: True
 
   # ---------------------------------------Optional Configuration-------------------------------- #
   ## -----------------------------------Configuration For Cost Critic--------------------------- ##
